scripts: formatted with shfmt
This was done because I'm very bored lol.

+++ scripts/build.sh
@@ -5,11 +5,11 @@
 
 BOLD="\033[1m"
 
-if [ -d "compiler/" ];then
-    :
+if [ -d "compiler/" ]; then
+	:
 else
-    echo -e "${BOLD}Compiler Not Found! Exitting..."
-    exit 0
+	echo -e "${BOLD}Compiler Not Found! Exitting..."
+	exit 0
 fi
 #if [ ! -d "$HOME/anykernel3/" ];then
 #    echo -e "${BOLD}Place Anykernel3 in ~/"
@@ -27,19 +27,19 @@ fi
 #    echo -e "${BOLD} Clang Not Installed!"
 #    exit 0
 #else
- #   :
+#   :
 #fi
-if ! hash ld.lld 2>/dev/null;then
-    echo -e "${BOLD} ld.lld Not Installed!"
-    exit 0
+if ! hash ld.lld 2>/dev/null; then
+	echo -e "${BOLD} ld.lld Not Installed!"
+	exit 0
 else
-    :
+	:
 fi
-if ! hash java 2>/dev/null;then
-    echo -e "${BOLD} java Not Installed!"
-    exit 0
+if ! hash java 2>/dev/null; then
+	echo -e "${BOLD} java Not Installed!"
+	exit 0
 else
-    :
+	:
 fi
 echo -e "${BOLD}
 ############o.O##############
@@ -56,12 +56,12 @@ echo -ne "${BOLD}Enter Device Codename: "
 read -r codename
 echo -ne "${BOLD}Enter Amount Of Cores To Use: "
 read -r core
-expr ${core} + 1 2> /dev/null
+expr ${core} + 1 2>/dev/null
 case $? in
-    0)
+0)
 	:
 	;;
-    *)
+*)
 	echo "error: Not a number"
 	exit 1
 	;;
@@ -80,49 +80,49 @@ MD5="$(md5sum NetErnels-"${codename}"-"${KERNELVER}"-signed.zip | cut -d' ' -f1)
 export MD5
 BUILDER="$(whoami)"
 export BUILDER
-tg_send_msg(){
-    curl -fsSL -X POST https://api.telegram.org/botBOTTOKEN/sendMessage -d "chat_id=-CHATID" -d "parse_mode=html" -d "text=${1}" &>/dev/null
+tg_send_msg() {
+	curl -fsSL -X POST https://api.telegram.org/botBOTTOKEN/sendMessage -d "chat_id=-CHATID" -d "parse_mode=html" -d "text=${1}" &>/dev/null
 }
 
-tg_send_build(){
-    curl -fsSL -X POST -F document=@"${1}" https://api.telegram.org/botBOTTOKEN/sendMessage \
-	 -F "chat_id=CHATID" \
-	 -F "disable_web_page_preview=true" \
-	 -F "parse_mode=html" \
-	 -F caption="${2}" &>/dev/null
+tg_send_build() {
+	curl -fsSL -X POST -F document=@"${1}" https://api.telegram.org/botBOTTOKEN/sendMessage \
+		-F "chat_id=CHATID" \
+		-F "disable_web_page_preview=true" \
+		-F "parse_mode=html" \
+		-F caption="${2}" &>/dev/null
 }
-zip_gen(){
-    tg_send_msg "<b>${BUILDER}</b>: <code>Zipping into a flashable zip ${codename}-${KERNELVER}</code>"
-    echo -e "${BOLD}Zipping into a flashable zip!"
-    mv ../compiled_code_"${codename}"/arch/arm64/boot/Image.gz ~/anykernel3/
-    zip -r NetErnels-"${codename}"-"${KERNELVER}".zip ~/anykernel3/*
-    cp NetErnels-"${codename}"-"${KERNELVER}".zip ./
-    echo -e "${BOLD}Preparing final zip!"
-    tg_send_msg "<b>${BUILDER}</b>: <code>Signing Zip file with AOSP keys ${codename}-${KERNELVER}</code>"
-    java -jar zipsigner.jar NetErnels-"${codename}"-"${KERNELVER}".zip NetErnels-"${codename}"-"${KERNELVER}"-signed.zip
-    tg_send_msg "<b>${BUILDER}</b>: <code>Build took : $((DIFF /60)) minute(s) and $((DIFF % 60)) second(s)</code>"
-    echo -e "${BOLD}Build took: $((DIFF /60)) minute(s) and $((DIFF % 60)) second(s)"
+zip_gen() {
+	tg_send_msg "<b>${BUILDER}</b>: <code>Zipping into a flashable zip ${codename}-${KERNELVER}</code>"
+	echo -e "${BOLD}Zipping into a flashable zip!"
+	mv ../compiled_code_"${codename}"/arch/arm64/boot/Image.gz ~/anykernel3/
+	zip -r NetErnels-"${codename}"-"${KERNELVER}".zip ~/anykernel3/*
+	cp NetErnels-"${codename}"-"${KERNELVER}".zip ./
+	echo -e "${BOLD}Preparing final zip!"
+	tg_send_msg "<b>${BUILDER}</b>: <code>Signing Zip file with AOSP keys ${codename}-${KERNELVER}</code>"
+	java -jar zipsigner.jar NetErnels-"${codename}"-"${KERNELVER}".zip NetErnels-"${codename}"-"${KERNELVER}"-signed.zip
+	tg_send_msg "<b>${BUILDER}</b>: <code>Build took : $((DIFF / 60)) minute(s) and $((DIFF % 60)) second(s)</code>"
+	echo -e "${BOLD}Build took: $((DIFF / 60)) minute(s) and $((DIFF % 60)) second(s)"
 }
 tg_send_msg "<b>${BUILDER}</b>: <code>Build Triggered! ${codename}-${KERNELVER}</code>"
 echo -ne "${BOLD}Do you want to do a clean build or a dirty build?(Y/n): "
 read -r build
 case ${build} in
-    Y|y)
+Y | y)
 	tg_send_msg "<b>${BUILDER}</b>: <code>Clean Build Commenced! ${codename}-${KERNELVER}</code>"
 	echo -e "${BOLD}Clean Build Commenced!"
 	make clean && make mrproper
-	if [ ! -d "../compiled_code_${codename}" ];then
-	    :
+	if [ ! -d "../compiled_code_${codename}" ]; then
+		:
 	else
-	    rm -rf ../compiled_code_"${codename}"
+		rm -rf ../compiled_code_"${codename}"
 	fi
 	;;
-    N|n)
+N | n)
 	tg_send_msg "<b>${BUILDER}</b>: <code>Dirty Build Commenced! ${codename}-${KERNELVER}</code>"
 	echo -e "${BOLD}Dirty Build Commenced!"
 	:
 	;;
-    *)
+*)
 	echo -e "${BOLD}Wrong Option Entered!"
 	exit 0
 	;;
@@ -142,15 +142,15 @@ make O=../compiled_code_"${codename}" -j${core} CC=clang 2>&1 | tee error.log
 BUILD_END=$(date +"%s")
 DIFF=$((BUILD_END - BUILD_START))
 
-if [ -f "${DIR}/../compiled_code_${codename}/arch/arm64/boot/Image.gz" ];then
-    tg_send_msg "<b>${BUILDER}</b>: <code>Kernel Successfully Compiled! ${codename}-${KERNELVER}</code>"
-    echo -e "${BOLD}Kernel Successfully Compiled..."
-    zip_gen
-    tg_send_build "NetErnels-${codename}-${KERNELVER}-signed.zip" "<b>${codename}-${KERNELVER}</b> | <b>MD5 Checksum</b>: <code>${MD5}</code>"
-    exit 0
+if [ -f "${DIR}/../compiled_code_${codename}/arch/arm64/boot/Image.gz" ]; then
+	tg_send_msg "<b>${BUILDER}</b>: <code>Kernel Successfully Compiled! ${codename}-${KERNELVER}</code>"
+	echo -e "${BOLD}Kernel Successfully Compiled..."
+	zip_gen
+	tg_send_build "NetErnels-${codename}-${KERNELVER}-signed.zip" "<b>${codename}-${KERNELVER}</b> | <b>MD5 Checksum</b>: <code>${MD5}</code>"
+	exit 0
 else
-    tg_send_msg "<b>${BUILDER}</b>: <code>Build failed to compile after $((DIFF / 60)) minutes and $((DIFF % 60)) seconds ${codename}-${KERNELVER}</code>"
-    echo -e "${BOLD}Build failed to compile..."
-    tg_send_build "error.log" "<b>Log of build</b>"
-    exit 0
+	tg_send_msg "<b>${BUILDER}</b>: <code>Build failed to compile after $((DIFF / 60)) minutes and $((DIFF % 60)) seconds ${codename}-${KERNELVER}</code>"
+	echo -e "${BOLD}Build failed to compile..."
+	tg_send_build "error.log" "<b>Log of build</b>"
+	exit 0
 fi
